fix: 移除频谱

+++ src-tauri/src/audio/analyzer_cmds.rs
@@ -6,7 +6,7 @@ use crate::error::AppResult;
 #[tauri::command]
 pub fn analyzer_start(
     state: tauri::State<'_, AudioState>,
-    on_data: Channel<Vec<u8>>,
+    on_data: Channel<Vec<f32>>,
 ) -> AppResult<()> {
     let s = state.lock().map_err(|e| e.to_string())?;
     s.analyzer.start(on_data);
